refactor(1013): extract GlobalWiredClientReportGenerator to src/reports/ (SC-001, position 36)
Cat B fresh extraction (SC-001 facade pattern). Moves the 251-line
GlobalWiredClientReportGenerator class from MistHelper.py to a new
module src/reports/global_wired_client_report_generator.py.

Live-global reads (apisession, ConfigUtils, FilterOperatorEngine,
InputUtils, DataProcessingUtils, DataExporter) are resolved via lazy
importlib.import_module("MistHelper") inside each helper that touches
them. Callers continue to reach the class through the
MistHelper.GlobalWiredClientReportGenerator re-export alias.

- New: src/reports/global_wired_client_report_generator.py (all 14
  static methods, identical behavior)
- MistHelper.py: class body deleted, replaced with breadcrumb;
  alphabetical re-export alias added
- pyproject.toml: coverage omit + Wave 2 mypy override for
  src.reports.*

Local gate: black clean, ruff clean, py_compile ok, MistHelper --test
baseline 66/0/131 preserved.
