+++ pyproject.toml
@@ -195,6 +195,7 @@ module = [
     "src.inventory.*",
     "src.org.*",
     "src.device.*",
+    "src.reports.*",
     "src.site.*",
     "src.troubleshooting.*",
     "src.websocket.*",
@@ -251,6 +252,7 @@ omit = [
     "src/gateway/gateway_ha_exporter.py",
     "src/input/prompt_client_utils.py",
     "src/inventory/org_device_inventory_summary_facade.py",
+    "src/reports/global_wired_client_report_generator.py",
     "src/reports/sfp_transceiver_data_processor.py",
     "src/reports/wired_client_manufacturer_report_generator.py",
     "src/site/bulk_radius_wlan_config_manager.py",
